+++ src/exceptions.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <print>
 
-void SyntaxError(std::string msg, uint32_t line, uint32_t col) {
+void SyntaxError(std::string msg, size_t line, size_t col) {
     std::println(
         std::cerr, "SyntaxError: {}\n\t at line {}, col {}", msg, line, col
     );
